refactor: move to utils folder
